feat: DailyFetchUseCase 정의
- 각 Entity의 UUID 옵셔널 제거
- DailyModel의 Int16 -> Int
- transform 메서드 생성해서 유즈ㅋ이스에서 변환할 수 있도록 구현

+++ Tempus/Domain/Model/DailyModel.swift
@@ -20,7 +20,7 @@ struct DailyModel {
         entity.uuid = id
         entity.title = title
         entity.startTime = startTime
-        entity.repeatCount = Int16(repeatCount)
+        entity.repeatCount = repeatCount
         entity.focusTime = focusTime
         entity.breakTime = breakTime
